+++ src/ops/utils/ExportImportUtils.test.js
@@ -1,12 +1,12 @@
-import {
-  convertArrayToBase64Script,
-  convertBase64ScriptToArray,
-  saveToFile,
-  validateImport,
-  checkTargetCompatibility
-} from '../../../src/ops/utils/ExportImportUtils'
 import { resolve } from 'path';
 import { rmSync, existsSync, readFileSync } from 'fs';
+import {
+  convertTextArrayToBase64,
+  convertBase64TextToArray,
+  saveToFile,
+  validateImport,
+  checkTargetCompatibility,
+} from './ExportImportUtils.js';
 // Warning! implimentation file contains non determinisitc functions which are either; not reasonable to test or imposible
 // Cause: date based non overidable functions
 // Not tested: getCurrentTimestamp
@@ -14,8 +14,7 @@ import { rmSync, existsSync, readFileSync } from 'fs';
 const FS_TMP_DIR = resolve('.', 'test', 'fs_tmp');
 const PATH_TO_ARTIFACT = resolve(FS_TMP_DIR, 'export.json');
 
-
-test('convertBase64ScriptToArray returns an array of script lines in base64 encoding', () => {
+test('convertBase64TextToArray returns an array of text lines in base64 encoding', () => {
   // Arrange
   const originalScript = `
     function frodo() {
@@ -24,12 +23,12 @@ test('convertBase64ScriptToArray returns an array of script lines in base64 enco
   `;
   const base64Script = Buffer.from(originalScript).toString('base64');
   // Act
-  const result = convertBase64ScriptToArray(base64Script);
+  const result = convertBase64TextToArray(base64Script);
   // Assert
   expect(result).toEqual(originalScript.split('\n'));
 });
 
-test('convertArrayToBase64Script', () => {
+test('convertTextArrayToBase64', () => {
   // Arrange
   const originalArrayOfScriptLines = `
     function frodo() {
@@ -38,24 +37,24 @@ test('convertArrayToBase64Script', () => {
   `;
   const expected = Buffer.from(originalArrayOfScriptLines).toString('base64');
   // Act
-  const result = convertArrayToBase64Script(originalArrayOfScriptLines.split('\n'));
+  const result = convertTextArrayToBase64(
+    originalArrayOfScriptLines.split('\n')
+  );
   // Assert
   expect(result).toEqual(expected);
 });
 
 test('validateImport should always return true', () => {
   expect(validateImport()).not.toBe(false);
-})
+});
 
 // This function has no way to determine when its asnyc task is complete, suggest using callback or promises to allow for testing
 describe.skip('file system based tests', () => {
-
   afterAll(() => {
     if (existsSync(PATH_TO_ARTIFACT)) {
       rmSync(PATH_TO_ARTIFACT);
     }
   });
-  
 
   test('saveToFile should save a file to specified tmp directory with expected data format', async () => {
     // Arrange
@@ -65,8 +64,8 @@ describe.skip('file system based tests', () => {
         id,
         location: 'The Shire',
         character: 'Gandalf',
-        words: 1064 
-      }
+        words: 1064,
+      },
     ];
 
     const expected = {
@@ -75,18 +74,17 @@ describe.skip('file system based tests', () => {
           id: 'id-3021',
           location: 'The Shire',
           character: 'Gandalf',
-          words: 1064
-        }
-      }
+          words: 1064,
+        },
+      },
     };
     // Act
-    // saveToFile('lotr', data, 'id', PATH_TO_ARTIFACT);
+    saveToFile('lotr', data, 'id', PATH_TO_ARTIFACT);
     const resultingJSON = JSON.parse(readFileSync(PATH_TO_ARTIFACT));
     // Assert
     expect(resultingJSON.lotr).toEqual(expected.lotr);
   });
 
-  
   test('saveToFile should save a file with metadata', async () => {
     // Arrange
     const id = `id-3021`;
@@ -95,8 +93,8 @@ describe.skip('file system based tests', () => {
         id,
         location: 'The Shire',
         character: 'Gandalf',
-        words: 1064 
-      }
+        words: 1064,
+      },
     ];
     // Act
     saveToFile('lotr', data, 'id', PATH_TO_ARTIFACT);
@@ -108,7 +106,7 @@ describe.skip('file system based tests', () => {
       'exportDate',
       'exportTool',
       'exportToolVersion',
-    ])
+    ]);
   });
 });
 
@@ -117,5 +115,5 @@ test.skip('checkTargetCompatibility to be made testable', () => {
   // Act
   const result = checkTargetCompatibility();
   // Assert
-  expect(result).toBe("not tested yet");
+  expect(result).toBe('not tested yet');
 });